update to commit [1.12#470b77e](https://github.com/GregTechCEu/GregTech/commits/470b77e640aa52b66654fb8a2d033c57f554c61b)

+++ src/main/java/com/gregtechceu/gtceu/common/pipelike/block/cable/CableBlock.java
@@ -119,7 +119,7 @@ public class CableBlock extends PipeMaterialBlock implements IBurnable {
                 EnergyFlowLogic logic = node.getData().getLogicEntryNullable(EnergyFlowLogic.TYPE);
 
                 if (logic != null) {
-                    long tick = GTCEu.getMinecraftServer().getTickCount();
+                    int tick = GTCEu.getMinecraftServer().getTickCount();
                     long cumulativeDamage = 0;
                     for (EnergyFlowData data : logic.getFlow(tick)) {
                         cumulativeDamage += (GTUtil.getTierByVoltage(data.voltage()) + 1) * data.amperage() * 4;
